[nextra] Mdast separate for fs dep
