+++ packages/aptos-nextra-components/src/utils/mdast/mdast.ts
@@ -1,23 +1,11 @@
-import fs from "node:fs";
-import path from "node:path";
 import { fromMarkdown } from "mdast-util-from-markdown";
 import { toMarkdown } from "mdast-util-to-markdown";
-import { Code, InlineCode, Root, Text } from "mdast-util-from-markdown/lib";
+import { Code, Root } from "mdast-util-from-markdown/lib";
 import { visit } from "unist-util-visit";
 
-/**
- * Read File as Tree
- */
-export function readFileAsTree(relativePath: string): Root {
-  const localPath = path.resolve(__dirname, relativePath);
-  const doc = fs.readFileSync(localPath);
-  const tree = fromMarkdown(doc);
+export function readMarkdownString(source: string): Root {
+  const tree = fromMarkdown(source);
   return tree;
-}
-
-export function writeFile(relativePath: string, data: string) {
-  const localPath = path.resolve(__dirname, relativePath);
-  fs.writeFileSync(localPath, data);
 }
 
 /**
